Revert "Revert "Switch to thin, versioned ES clients (#5217)""
This reverts commit cdab452fa9999762d78d612693ad0c07b28a84c1.
